+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Added
 * Added Troubleshooting information about Crawler Module
+* Added PageService
 
 ### Changed
 * Switch to Constants for Repository TableNames
+* Use ConfigurationRepository function instead of queries in CrawlerController
 * Following classes are marked as internal
   * Backend/BackendModule.php
   * Backend/Helper/ResultHandler.php
@@ -47,6 +49,7 @@
 ### Deprecated
 #### Classes
 #### Functions & Properties
+* CrawlerController->checkIfPageShouldBeSkipped()
 
 ### Removed
 #### Classes